Add config; Add multi-instances
TODO: Check out the initialization of TP and PP. Now each process
actually has information of all the groups, even for which this process
does not belong to. Otherwise the torch.distributed.broadcast may be
blocked. I am not sure whether this will become a scalability bottleneck
or not in the future.

TODO: Check whether the model loading is correct or not.